Simplify the implementation of `zip`
# Motivation
We recently merged the PR that corrected the `Sendable` constraint of the. After thinking about this some more, I realised we overly complicated the implementation here. The great thing about `zip` is that it is actually trivial and we don't have to race any of the upstream sequences against each other. As soon as I realised that, I knew we can completely get rid of the `Task` and just implement this on top of a task group.

# Modification
Remove the complicated state machine implementation and replace it with a straight forward task group.

# Result
This makes the implementation of `zip` tremendously shorter. Furthermore, it improves throughout by roughly 1.5 times.