Make `main` build with 5.7 and 5.8 again
# Motivation
In one of the recent commits, we have broken the build on Swift 5.7 and 5.8.

# Modification
This PR fixed the compiler errors and makes us build on 5.7 and 5.8 again. I introduced a bunch of availability to the validation setup since the `SerialExecutor` is otherwise not implementable without a warning.

# Result
Works on all of our supported Swift versions again.