Add support for the Swift package index
# Motivation
We had some scripts around to generate docs for this package but all of our other packages are now providing documentation on the Swift package index.

# Modification
This removes the old script to generate documentation and adds an `.spi.yml` file. Furthermore, I removed the conditional dependency on the docc plugin which we are unconditionally depending on in most of our other packages like NIO.